+++ Tests/MongoDb.MultiTenancy.SeperateDb/MongoDb.MultiTenancy.SeperateDb.Api/Startup.cs
@@ -1,0 +1,73 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using MongoDb.MultiTenancy.SeperateDb.Api.Configuration;
+using MongoDb.MultiTenancy.SeperateDb.Api.Filters;
+using MongoDb.MultiTenancy.SeperateDb.Application;
+using MongoDb.MultiTenancy.SeperateDb.Infrastructure;
+using Serilog;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.AspNetCore.Startup", Version = "1.0")]
+
+namespace MongoDb.MultiTenancy.SeperateDb.Api
+{
+    [IntentManaged(Mode.Merge)]
+    public class Startup
+    {
+        public Startup(IConfiguration configuration)
+        {
+            Configuration = configuration;
+        }
+
+        public IConfiguration Configuration { get; }
+
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddControllers(
+                opt =>
+                {
+                    opt.Filters.Add<ExceptionFilter>();
+                });
+            services.AddApplication(Configuration);
+            services.ConfigureApplicationSecurity(Configuration);
+            services.ConfigureHealthChecks(Configuration);
+            services.ConfigureProblemDetails();
+            services.ConfigureApiVersioning();
+            services.AddInfrastructure(Configuration);
+            services.ConfigureSwagger(Configuration);
+            services.ConfigureMultiTenancy(Configuration);
+        }
+
+        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+            }
+            app.UseSerilogRequestLogging();
+            app.UseExceptionHandler();
+            app.UseHttpsRedirection();
+            app.UseRouting();
+            app.UseMultiTenancy();
+            app.UseAuthentication();
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapDefaultHealthChecks();
+                endpoints.MapControllers();
+            });
+            app.UseSwashbuckle(Configuration);
+        }
+    }
+}