+++ Tests/MongoDb.MultiTenancy.SeperateDb/MongoDb.MultiTenancy.SeperateDb.Application/Customers/DeleteCustomer/DeleteCustomerCommand.cs
@@ -1,0 +1,19 @@
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+using MongoDb.MultiTenancy.SeperateDb.Application.Common.Interfaces;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandModels", Version = "1.0")]
+
+namespace MongoDb.MultiTenancy.SeperateDb.Application.Customers.DeleteCustomer
+{
+    public class DeleteCustomerCommand : IRequest, ICommand
+    {
+        public DeleteCustomerCommand(string id)
+        {
+            Id = id;
+        }
+
+        public string Id { get; set; }
+    }
+}