+++ Tests/MongoDb.MultiTenancy.SeperateDb/MongoDb.MultiTenancy.SeperateDb.Application/Customers/GetCustomers/GetCustomersQuery.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+using MongoDb.MultiTenancy.SeperateDb.Application.Common.Interfaces;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.QueryModels", Version = "1.0")]
+
+namespace MongoDb.MultiTenancy.SeperateDb.Application.Customers.GetCustomers
+{
+    public class GetCustomersQuery : IRequest<List<CustomerDto>>, IQuery
+    {
+        public GetCustomersQuery()
+        {
+        }
+    }
+}