+++ Tests/MongoDb.MultiTenancy.SeperateDb/MongoDb.MultiTenancy.SeperateDb.Infrastructure/Repositories/CustomerMongoRepository.cs
@@ -1,0 +1,32 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using MongoDb.MultiTenancy.SeperateDb.Domain.Entities;
+using MongoDb.MultiTenancy.SeperateDb.Domain.Repositories;
+using MongoDb.MultiTenancy.SeperateDb.Infrastructure.Persistence;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.MongoDb.Repositories.Repository", Version = "1.0")]
+
+namespace MongoDb.MultiTenancy.SeperateDb.Infrastructure.Repositories
+{
+    public class CustomerMongoRepository : MongoRepositoryBase<Customer>, ICustomerRepository
+    {
+        public CustomerMongoRepository(ApplicationMongoDbContext context) : base(context)
+        {
+        }
+
+        public async Task<Customer?> FindByIdAsync(string id, CancellationToken cancellationToken = default)
+        {
+            return await FindAsync(x => x.Id == id, cancellationToken);
+        }
+
+        public async Task<List<Customer>> FindByIdsAsync(string[] ids, CancellationToken cancellationToken = default)
+        {
+            return await FindAllAsync(x => ids.Contains(x.Id), cancellationToken);
+        }
+    }
+}